ping_server/requirements.txt
fastapi
uvicorn
pydantic
pydantic_settings
python-dotenv